Replaced the use of floats with MoneyPHP.

+++ src/contracts/Coinverter.php
@@ -10,9 +10,9 @@ interface Coinverter
 
     public function exchangeRateBetweenDateRange(string $from, string $to, Carbon $date, Carbon $endDate);
 
-    public function convert(float $amount, string $from, string $to, Carbon $date = null);
+    public function convert(int $amount, string $from, string $to, Carbon $date = null);
 
-    public function convertBetweenDateRange(float $amount, string $from, string $to, Carbon $date, Carbon $endDate);
+    public function convertBetweenDateRange(int $amount, string $from, string $to, Carbon $date, Carbon $endDate);
 
     public function currencies();
 }
